Bugfix: accessing graph.idiograph.allograph on a new Graph was raising an exception (silently) each time someone tried to save a new graph using the annotator. Fixed  by checking if the graph is new. See annotation::save()

+++ views/annotation.py
@@ -506,7 +506,8 @@ def save(request, graphs):
 
                         scribe = hand.scribe
 
-                        if allograph != graph.idiograph.allograph.id:
+                        # GN: if this is a new Graph, it has no idiograph yet, so we test this first
+                        if graph.id and (allograph != graph.idiograph.allograph.id):
                             graph.graph_components.all().delete()
 
                         idiograph_list = Idiograph.objects.filter(allograph=allograph,
